src/endpoint/lol/spells.rs
use std::future::Future;

use reqwest::{Request, Response};
use url::Url;

use crate::{
    endpoint::{sealed::Sealed, CollectionOptions, EndpointError, ListResponse, BASE_URL},
    model::lol::spell::Spell,
};

#[derive(Debug, Clone, Eq, PartialEq, Default)]
pub struct ListSpells(pub CollectionOptions);

impl Sealed for ListSpells {
    type Response = ListResponse<Spell>;

    fn to_request(self) -> Result<Request, EndpointError> {
        let mut url = Url::parse(&format!("{}/lol/spells", BASE_URL))?;
        self.0.add_params(&mut url);
        Ok(Request::new(reqwest::Method::GET, url))
    }

    fn from_response(
        response: Response,
    ) -> impl Future<Output = Result<Self::Response, EndpointError>> + Send {
        ListResponse::from_response(response)
    }
}
